# app/static/swagger.yaml
swagger: '2.0'
info:
  title: Mechanic Shop API
  description: REST API for customers, mechanics, inventory, and service tickets.
  version: 1.0.0
host: "mechanic-shop-api-4ig7.onrender.com"
basePath: /
schemes:
  - "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'Use format: Bearer <token>'
paths:
...
  /service_tickets:
    post:
      tags:
        - Service Tickets
      summary: Create a service ticket
      description: Creates a new service ticket linked to a customer.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateServiceTicketPayload'
      responses:
        201:
          description: Service ticket created
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: []
              inventory_ids: []
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        500:
          description: Failed to create service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to create service ticket.
    get:
      tags:
        - Service Tickets
      summary: List service tickets
      description: Returns a paginated list of service tickets. If pagination values are invalid, all tickets are returned.
      parameters:
        - in: query
          name: page
          type: integer
          required: false
          default: 1
        - in: query
          name: per_page
          type: integer
          required: false
          default: 10
      responses:
        200:
          description: Service ticket list
          schema:
            type: array
            items:
              $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              - id: 1
                vin: 1HGCM82633A123456
                service_date: '2023-08-15'
                description: Brake pads replacement
                status: Pending
                customer_id: 1
                mechanic_ids: [1]
                inventory_ids: [1]
  /service_tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - Service Tickets
      summary: Assign a mechanic to a ticket
      description: Assigns the specified mechanic to a service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: path
          name: mechanic_id
          required: true
          type: integer
      responses:
        200:
          description: Ticket updated
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: [1]
              inventory_ids: [1]
        404:
          description: Ticket or mechanic not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Mechanic not found.
        500:
          description: Failed to update service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update service ticket.
  /service_tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - Service Tickets
      summary: Remove a mechanic from a ticket
      description: Removes the specified mechanic from a service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: path
          name: mechanic_id
          required: true
          type: integer
      responses:
        200:
          description: Ticket updated
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: []
              inventory_ids: [1]
        400:
          description: Mechanic not assigned to ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Mechanic is not assigned to this ticket.
        404:
          description: Ticket or mechanic not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Service ticket not found.
        500:
          description: Failed to update service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update service ticket.
  /service_tickets/{ticket_id}/assign-customer/{customer_id}:
    put:
      tags:
        - Service Tickets
      summary: Assign a customer to a ticket
      description: Reassigns a service ticket to the specified customer.
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: path
          name: customer_id
          required: true
          type: integer
      responses:
        200:
          description: Ticket updated
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: [1]
              inventory_ids: [1]
        404:
          description: Ticket or customer not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Customer not found.
        500:
          description: Failed to update service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update service ticket.
  /service_tickets/{ticket_id}/add-part/{inventory_id}:
    put:
      tags:
        - Service Tickets
      summary: Add one part to a ticket
      description: Adds one inventory part to a service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: path
          name: inventory_id
          required: true
          type: integer
      responses:
        200:
          description: Ticket updated
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: [1]
              inventory_ids: [1]
        404:
          description: Ticket or inventory item not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Inventory item not found.
        500:
          description: Failed to update service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update service ticket.
  /service_tickets/{ticket_id}/edit:
    put:
      tags:
        - Service Tickets
      summary: Add and remove mechanics on a ticket in one request
      description: Adds and removes mechanics on one service ticket in a single update payload.
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/EditTicketMechanicsPayload'
      responses:
        200:
          description: Ticket updated
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: [1]
              inventory_ids: [1]
        404:
          description: Ticket or mechanic not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Mechanic 9999 not found.
        500:
          description: Failed to update service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update service ticket.
  /service_tickets/{ticket_id}/edit-parts:
    put:
      tags:
        - Service Tickets
      summary: Add and remove parts on a ticket in one request
      description: Adds and removes inventory parts on one service ticket in a single update payload.
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/EditTicketPartsPayload'
      responses:
        200:
          description: Ticket updated
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 1
              vin: 1HGCM82633A123456
              service_date: '2023-08-15'
              description: Brake pads replacement
              status: Pending
              customer_id: 1
              mechanic_ids: [1]
              inventory_ids: [1]
        404:
          description: Ticket or inventory item not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Inventory item 9999 not found.
        500:
          description: Failed to update service ticket
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update service ticket.
definitions:
...
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
        example: 1
      vin:
        type: string
        example: 1HGCM82633A123456
      service_date:
        type: string
        format: date
        example: '2023-08-15'
      description:
        type: string
        example: Brake pads replacement
      status:
        type: string
        example: Pending
      customer_id:
        type: integer
        example: 1
      mechanic_ids:
        type: array
        items:
          type: integer
        example: [1]
      inventory_ids:
        type: array
        items:
          type: integer
        example: [1, 2]
...
  CreateServiceTicketPayload:
    type: object
    properties:
      vin:
        type: string
        example: 1HGCM82633A123456
      service_date:
        type: string
        format: date
        example: '2023-08-15'
      description:
        type: string
        example: Brake pads replacement
      status:
        type: string
        example: Pending
      customer_id:
        type: integer
        example: 1
    required:
      - vin
      - service_date
      - description
      - status
      - customer_id
  EditTicketMechanicsPayload:
    type: object
    properties:
      add_ids:
        type: array
        items:
          type: integer
        example: [2]
      remove_ids:
        type: array
        items:
          type: integer
        example: [4]
  EditTicketPartsPayload:
    type: object
    properties:
      add_part_ids:
        type: array
        items:
          type: integer
        example: [3]
      remove_part_ids:
        type: array
        items:
          type: integer
        example: [7]
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: Validation failed for one or more fields.